Improvements in the preprocessing and in the callback.
The constraints are now sorted and the duplicates removed.
The ones that are found more than once are brought at the top of the list.

In the callback now there are no memory allocations. The computation for the fractionality and the estimate of the scores are separated. This reduces part of the computation time inside the callbacks by  applying a look ahead policy.